929-UniqueEmailAddresses_test.go
package main

import "testing"

func Test_numUniqueEmails(t *testing.T) {
	type args struct {
		emails []string
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{"Example 1", args{[]string{"test.email+alex@leetcode.com", "test.e.mail+bob.cathy@leetcode.com", "testemail+david@lee.tcode.com"}}, 2},
		{"Same", args{[]string{"test@gmail.com", "test@gmail.com", "test@gmail.com"}}, 1},
		{"Repeat", args{[]string{"test+@gmail.com", "test.@gmail.com", "test@gmail.com"}}, 1},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := numUniqueEmails(tt.args.emails); got != tt.want {
				t.Errorf("numUniqueEmails() = %v, want %v", got, tt.want)
			}
		})
	}
}
